compiled project one and two

+++ doubly_linked_lists/0-print_dlistint.c
@@ -1,12 +1,12 @@
 #include "lists.h"
 /**
 * print_dlistint - a function that prints all the elements of a dlistint_t list
-* @h : pointer constant the dlistint
+* @h: pointer constant the dlistint
 * Return: the number of node
 */
 size_t print_dlistint(const dlistint_t *h)
 {
-int data = 0;
+size_t data = 0;
 if (h == NULL)
 {
 return (0);
